+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Fix amending layout on iOS
+  - https://github.com/Shopify/flash-list/pull/412
 - Define `FlashList` props previously inherited from `VirtualizedList` and `FlatList` explicitly
   - https://github.com/Shopify/flash-list/pull/386
 - Make `estimatedItemSize` optional
   - https://github.com/Shopify/flash-list/pull/378
+- Change `overrideItemType` prop name to `getItemType`
+  - https://github.com/Shopify/flash-list/pull/369
+- Added `CellRendererComponent` prop
+  - https://github.com/Shopify/flash-list/pull/362
+- Added automatic height measurement for horizontal lists even when parent isn't deterministic
+  - https://github.com/Shopify/flash-list/pull/409
 
 ## [0.5.0] - 2022-04-29